move button back to the resources
